+++ CHANGELOG.md
@@ -26,7 +26,10 @@ and this project aspires to adhere to [Semantic Versioning](https://semver.org/s
 - Added support for 64 bit IDs
 - Added support for RZ meshes
 - If a `default_dir` is passed as an Ascent Option, extracts and plots will output to that directory by default now.
-- Added a warning if a `ascent.py` file exists in the current working directory when executing a python extract.
+- Added a warning if a `ascent.py` file exsits in the current working directory when executing a python extract. 
+- Added support for monochrome/solid color plotting
+- Added support for `surface` and `wireframe` aliases
+- Added support for plotting pseudocolor plots without providing a field
 
 ### Changed
 - Redefined the parameter verification and surprise checking behavior for Filters to instead use a conduit based JSON Schema style verification schema. Implemented for all existing filters.